Refactor implementation of expiration handling use cases using useEffect hook
The useEffect hook is used to subscribe to events and activate time interval driven function when the component mounts, and unsubscribe to those events and clear time interval when the component unmounts. As a result, the useEffect hook eliminates the need to subscribe to the window's 'load' event, as well as helps to avoid memory leak.