Removed string and text transformation as they are now done in server-side; updated prop for 'fields' search param
Also, added a non-display column for gene_symbol in react-table to fix a bug in which time series plot header showing multiple gene ids if user searches multiple genes. The cause was due to the <a /> transformation of gene_symbol and the Redux 'geneSymbol' state was used (which supports one or more gene ids).